+++ window_scene/screen_session_manager/src/fold_screen_controller/single_display_pocket_fold_policy.cpp
@@ -107,8 +107,8 @@ void SingleDisplayPocketFoldPolicy::GetFoldCreaseRect(bool isVertical,
         TLOGI(WmsLogTag::DMS, "the current FoldCreaseRect is horizontal");
         liveCreaseRegionPosX = foldRect[0];
         liveCreaseRegionPosY = foldRect[1];
-        liveCreaseRegionPosWidth = foldRect[2];
-        liveCreaseRegionPosHeight = foldRect[3];
+        liveCreaseRegionPosWidth = static_cast<uint32_t>(foldRect[2]);
+        liveCreaseRegionPosHeight = static_cast<uint32_t>(foldRect[3]);
     }
     foldCreaseRect = {
         {
